déplacement de la classe enquete dans un fichier apart
