+++ utilitaire/utils.py
